Address comments - 1
